Avoid some errors and warnings from clang in parse_integer_literal
In parse_integer_literal, we index into an array of digits at locations 0 and 1 in a constexpr context. These lookups are guarded by a check on the length of the array.

Even though these lookups are guarded, clang complains they are at locations that do not exist when the length of the array is small.

This change silences clang's complaints by introducing some padding into the array of digits to ensure that locations 0 and 1 always exist.